+++ src/libstore/include/nix/store/references.hh
