+++ src/libutil/meson.build
@@ -145,7 +145,6 @@ sources = [ config_priv_h ] + files(
   'pos-table.cc',
   'position.cc',
   'posix-source-accessor.cc',
-  'references.cc',
   'serialise.cc',
   'signature/local-keys.cc',
   'signature/signer.cc',
